+++ lib/handle.rs
@@ -8,7 +8,7 @@ use crate::Middleware;
 use std::panic;
 
 // Import local files
-use crate::common::{any_string, reason_phrase, trim_end_bytes};
+use crate::common::{any_string, has_header, reason_phrase, trim_end_bytes};
 use crate::content_type::Content;
 use crate::header::{headers_to_string, Header};
 use crate::http;
@@ -206,11 +206,19 @@ pub(crate) fn response_http(
     };
 
     // Add default headers to response
-    let mut headers = res.headers.to_owned();
-    headers.append(&mut this.default_headers.to_owned());
+    // Only the ones that arent already in the response
+    let mut headers = res.headers.clone();
+    for i in &this.default_headers {
+        if !has_header(&headers, &i.name) {
+            headers.push(i.clone());
+        }
+    }
+    // headers.append(&mut this.default_headers.clone());
 
-    // Add content-length header to response
-    headers.push(Header::new("Content-Length", &res.data.len().to_string()));
+    // Add content-length header to response if it hasent already been deifned by the route or defult headers
+    if !has_header(&headers, "Content-Length") {
+        headers.push(Header::new("Content-Length", &res.data.len().to_string()));
+    }
 
     // Convert the response to a string
     // TODO: Use Bytes instead of String